requirements.txt
Flask==2.3.3
werkzeug==2.3.7
python-dotenv
google-generativeai
pypdf
python-docx